✨ Feat: 색상 선택 구현
Related to: #350

+++ app/src/main/java/com/mongmong/namo/data/datasource/auth/RemoteAuthDataSource.kt
@@ -42,7 +42,7 @@ class RemoteAuthDataSource @Inject constructor(
         name: String,
         nickname: String,
         birthday: String,
-        colorId: Long,
+        colorId: Int,
         bio: String,
         profileImage: String
     ): Boolean {
