+++ app/src/main/java/com/mongmong/namo/data/repositoriyImpl/AuthRepositoryImpl.kt
@@ -19,7 +19,7 @@ class AuthRepositoryImpl @Inject constructor(
         name: String,
         nickname: String,
         birthday: String,
-        colorId: Long,
+        colorId: Int,
         bio: String,
         profileImage: String
     ): Boolean {
